requirements.txt
pylint==3.0.3

python-dotenv==1.0.0
gitpython
streamlit
GitPython
